// src/actions/types.ts
/* eslint-disable import/prefer-default-export */
export enum Types {
  FETCH_ENTRIES_STARTED = 'FETCH_ENTRIES_STARTED',
  FETCH_ENTRIES_DONE = 'FETCH_ENTRIES_DONE',
  FETCH_ENTRIES_FAILED = 'FETCH_ENTRIES_FAILED',
  HEADER_TITLE_CHANGED = 'HEADER_TITLE_CHANGED',
  HEADER_HIDE_BACK_BUTTON = 'HEADER_HIDE_BACK_BUTTON',
  HEADER_DISPLAY_BACK_BUTTON = 'HEADER_DISPLAY_BACK_BUTTON',
  FETCH_THUMBNAIL_STARTED = 'FETCH_THUMBNAIL_STARTED',
  FETCH_THUMBNAIL_DONE = 'FETCH_THUMBNAIL_DONE',
  FETCH_THUMBNAIL_FAILED = 'FETCH_THUMBNAIL_FAILED',
  FILE_DIALOG_OPEN = 'FILE_DIALOG_OPEN',
  FILE_DIALOG_CLOSE = 'FILE_DIALOG_CLOSE',
  FILE_DIALOG_HIDE_APP_BAR = 'FILE_DIALOG_HIDE_APP_BAR',
  FILE_DIALOG_DISPLAY_APP_BAR = 'FILE_DIALOG_DISPLAY_APP_BAR',
  FILE_DIALOG_FILE_CHANGED = 'FILE_DIALOG_FILE_CHANGED',
  FETCH_IMAGES_STARTED = 'FETCH_IMAGES_STARTED',
  FETCH_IMAGES_DONE = 'FETCH_IMAGES_DONE',
  FETCH_IMAGE_DONE = 'FETCH_IMAGE_DONE',
  FETCH_IMAGES_FAILED = 'FETCH_IMAGES_FAILED',
  DISPLAY_NEXT_PAGE = 'DISPLAY_NEXT_PAGE',
  DISPLAY_PREV_PAGE = 'DISPLAY_PREV_PAGE',
  ADD_FILE_TO_FAVORITE = 'ADD_FILE_TO_FAVORITE',
  ADD_DIRECTORY_TO_FAVORITE = 'ADD_DIRECTORY_TO_FAVORITE',
  REMOVE_FROM_FAVORITE = 'REMOVE_FROM_FAVORITE',
  VIEWER_PER_PAGE_CHANGED = 'VIEWER_PER_PAGE_CHANGED',
  DELETE_FILE_STARTED = 'DELETE_FILE_STARTED',
  DELETE_FILE_DONE = 'DELETE_FILE_DONE',
  DELETE_FILE_FAILED = 'DELETE_FILE_FAILED',
}